% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhodes.R
\name{rhodes.curve}
\alias{rhodes.curve}
\title{Logistic Curve Fitting by Rhodes Method}
\usage{
rhodes.curve(p)
}
\arguments{
\item{p}{a numeric vector}
}
\value{
r.hat , L.hat , beta.hat   :    the estimated values of the parameters r, L and beta.

predicted.values  :   the predicted values of p
}
\description{
This function  fits the Logistic Curve in population Data by Rhodes Method along with estimates of the parameters and predicted value.
}
\details{
Suppose we have n observations from population size corresponding to n equivalent time points say, at t=0,1,...,n-1. Here we assume the Logistic law of population growth , p=L/(1+exp(r*(beta-t))).
}
\examples{
 u=c(12,15,16,18,16,21,25,27,29,30,35,36)
 rhodes.curve(u)
}
\author{
Arnab Roy, Debarghya Baul.
}
